Update react-i18next.js
The changes to dynamically load language resources in your React application would typically be made in the initialization code for react-i18next. This code may be in a separate module or file, or it may be included in your main App.js file.

If your initialization code is in a separate module or file, you would need to modify that file to replace the code that loads all language resources upfront with code that loads language resources on demand.

If your initialization code is in your App.js file, you would need to modify that file to remove the current code that loads all language resources upfront and replace it with code that loads language resources on demand, as described in my previous response.